EPMLSTRTAI-362 AdminControllerSecurityTest + AdminControllerTest *
